% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_quasi_mmtr.R
\name{bic_quasi_mmtr}
\alias{bic_quasi_mmtr}
\title{This function computes the EBIC for an MMTR model.}
\usage{
bic_quasi_mmtr(mmtr_mod, y, X, Z, group_ids, n_proc = 1)
}
\arguments{
\item{mmtr_mod}{MMTR model object produced by calling the mmtr() function.}

\item{y}{Numeric M-vector of responses across all N groups, where M = M1 + M2 + ... + MN.}

\item{X}{(D+1)-mode fixed effects tensor of dimension P1xP2x...xPDxM, i.e. there is a D-mode
P1xP2x...xPD fixed effects covariate tensor associated with each of the M responses.}

\item{Z}{(D+1)-mode random effects tensor of dimension Q1xQ2x...xQDxM, i.e. there is a D-mode
Q1xQ2x...xQD random effects covariate tensor associated with each of the M responses.}

\item{group_ids}{Integer or factor M-vector indicating which group each observation belongs to.}

\item{n_proc}{Number of processors to use for parallel computation. Defaults to one.}
}
\description{
This function computes the EBIC for an MMTR model.
}
