% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebic_mmtr.R
\name{ebic_mmtr}
\alias{ebic_mmtr}
\title{This function computes the EBIC for an MMTR model.}
\usage{
ebic_mmtr(
  mmtr_mod,
  y,
  X,
  Z,
  group_ids,
  n_reps = 1,
  conv_thrsh = 0.001,
  n_proc = 1
)
}
\arguments{
\item{mmtr_mod}{MMTR model object produced by calling the mmtr() function.}

\item{y}{Numeric M-vector of responses across all N groups, where M = M1 + M2 + ... + MN.}

\item{X}{(D+1)-mode fixed effects tensor of dimension P1xP2x...xPDxM, i.e. there is a D-mode
P1xP2x...xPD fixed effects covariate tensor associated with each of the M responses.}

\item{Z}{(D+1)-mode random effects tensor of dimension Q1xQ2x...xQDxM, i.e. there is a D-mode
Q1xQ2x...xQD random effects covariate tensor associated with each of the M responses.}

\item{group_ids}{Integer or factor M-vector indicating which group each observation belongs to.}

\item{n_reps}{Maximum number of iterations to run the mixed model trace regression AECM algorithm.
Defaults to 1000.}

\item{conv_thrsh}{Convergence threshold used for early stopping of the mixed model trace regression
AECM algorithm based on the change in the log-likelihood from one iteration to the next.
Defaults to 1e-3.}

\item{n_proc}{Number of processors to use for parallel computation. Defaults to one.}
}
\description{
This function computes the EBIC for an MMTR model.
}
