% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_sweep_mmtr.R
\name{lambda_sweep_mmtr}
\alias{lambda_sweep_mmtr}
\title{This function performs a mixed model trace regression (MMTR) lambda sweep.}
\usage{
lambda_sweep_mmtr(
  y,
  X,
  Z,
  group_ids,
  s_dims_max,
  n_reps = 1000,
  B_lambda_range = c(1e-04, 1),
  L_lambda_range = c(1e-04, 1),
  n_B_lambdas = 10,
  n_L_lambdas = 10,
  comp_ebic = TRUE,
  conv_thrsh = 0.001,
  n_proc = 1
)
}
\arguments{
\item{y}{Numeric M-vector of responses across all N groups, where M = M1 + M2 + ... + MN.}

\item{X}{(D+1)-mode fixed effects tensor of dimension P1xP2x...xPDxM, i.e. there is a D-mode
P1xP2x...xPD fixed effects covariate tensor associated with each of the M responses.}

\item{Z}{(D+1)-mode random effects tensor of dimension Q1xQ2x...xQDxM, i.e. there is a D-mode
Q1xQ2x...xQD random effects covariate tensor associated with each of the M responses.}

\item{group_ids}{Integer or factor M-vector indicating which group each observation belongs to.}

\item{s_dims_max}{Integer D-vector indicating the upper bound of the estimated rank of each mode's
random effects covariance matrix.}

\item{n_reps}{Maximum number of iterations to run the mixed model trace regression AECM algorithm
for each combination of lambda penalty parameters. Defaults to 1000.}

\item{B_lambda_range}{Minimum and maximum lambda penalty parameters indicating the range of
lambda values over which a parameter sweep will be performed in the estimation of the fixed
effects mean structure B. Defaults to (1e-4, 1).}

\item{L_lambda_range}{Minimum and maximum lambda penalty parameters indicating the range of
lambda values over which a parameter sweep will be performed in the estimation of the random
effects covariance matrices. Defaults to (1e-4, 1).}

\item{n_B_lambdas}{Number of different fixed effects lambda penalty parameters to use in lambda
sweep. Defaults to ten.}

\item{n_L_lambdas}{Number of different random effects lambda penalty parameters to use in lambda
sweep. Defaults to ten.}

\item{comp_ebic}{Logical value indicating whether or not the EBIC should be computed after the
MMTR model is fit for each combination of lambda penalty parameters. Defaults to TRUE.}

\item{conv_thrsh}{Convergence threshold used for early stopping of the mixed model trace regression
AECM algorithm based on the change in the objective function from one iteration to the next for
each combination of lambda penalty parameters. Defaults to 1e-3.}

\item{n_proc:}{Number of processors to use for parallel computation. Defaults to one.}
}
\description{
This function performs a mixed model trace regression (MMTR) lambda sweep.
}
