% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmtr.R
\name{mmtr}
\alias{mmtr}
\title{This function fits a mixed model trace regression (MMTR) model to matrix covariate data.}
\usage{
mmtr(
  y,
  X,
  Z,
  group_ids,
  L_list = NA,
  lambdas = c(1e-04, 1e-04),
  comp_ebic = TRUE,
  n_reps = 1000,
  conv_thrsh = 0.001,
  n_proc = 1
)
}
\arguments{
\item{y}{Numeric M-vector of responses across all N groups, where M = M1 + M2 + ... + MN.}

\item{X}{(D+1)-mode fixed effects tensor of dimension P1xP2x...xPDxM, i.e. there is a D-mode
P1xP2x...xPD fixed effects covariate tensor associated with each of the M responses.}

\item{Z}{(D+1)-mode random effects tensor of dimension Q1xQ2x...xQDxM, i.e. there is a D-mode
Q1xQ2x...xQD random effects covariate tensor associated with each of the M responses.}

\item{group_ids}{Integer or factor M-vector indicating which group each observation belongs to.}

\item{L_list}{List of D square root matrix components of the overall random effects covariance
matrix, which is just the squared Kronecker product of these component matrices in reverse
order. E.g. for D = 2, L_list[\link{1}] = L1 would be the square root of the column fibres'
covariance matrix, L_list[\link{2}] = L2 would be the square root of the row fibres' covariance
matrix, and thus L = L2 (x) L1 would be the square root of the overall random effects
covariance matrix to the vectorization of any group's random effects tensor. Defaults to NA.}

\item{lambdas}{Vector of lambda penalty parameters, the first of which pertains to the fixed
effects mean structure and the second of which pertains to the random effects covariance
matrices. Defaults to 1e-4 for each lambda.}

\item{comp_ebic}{Logical value indicating whether or not the EBIC should be computed after the MMTR
model is fit. Defaults to TRUE.}

\item{n_reps}{Maximum number of iterations to run the mixed model trace regression AECM algorithm.
Defaults to 1000.}

\item{conv_thrsh}{Convergence threshold used for early stopping of the mixed model trace regression
AECM algorithm based on the change in the objective function from one iteration to the next.
Defaults to 1e-3.}

\item{n_proc}{Number of processors to use for parallel computation. Defaults to one.}
}
\description{
This function fits a mixed model trace regression (MMTR) model to matrix covariate data.
}
