% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mmtr.R
\name{predict_mmtr}
\alias{predict_mmtr}
\title{This function computes predictions for an MMTR model.}
\usage{
predict_mmtr(mmtr_mod, X, Z, group_ids)
}
\arguments{
\item{mmtr_mod}{MMTR model object produced by calling the mmtr() function.}

\item{X}{(D+1)-mode fixed effects tensor of dimension P1xP2x...xPDxM, i.e. there is a D-mode
P1xP2x...xPD fixed effects covariate tensor associated with each of the M responses.}

\item{Z}{(D+1)-mode random effects tensor of dimension Q1xQ2x...xQDxM, i.e. there is a D-mode
Q1xQ2x...xQD random effects covariate tensor associated with each of the M responses.}

\item{group_ids}{Integer or factor M-vector indicating which group each observation belongs to.}
}
\description{
This function computes predictions for an MMTR model.
}
